Finish Desafio

+++ src/main/java/com/books/libreria/model/DataLibros.java
@@ -7,9 +7,9 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record DataLibros(
-                @JsonAlias("title") String titulo,
-                @JsonAlias("authors") List<DataAutor> autor,
-                @JsonAlias("languages") List<String> idioma,
-                @JsonAlias("download_count") Integer descargas) {
+        @JsonAlias("title") String titulo,
+        @JsonAlias("authors") List<DataAutor> autor,
+        @JsonAlias("languages") List<String> idioma,
+        @JsonAlias("download_count") Double descargas) {
 
 }
